+++ src/sys/classes/st/impls/shell/ftn-custom/zshell.c
@@ -12,20 +12,23 @@
 #include <slepcst.h>
 
 #if defined(PETSC_HAVE_FORTRAN_CAPS)
-#define stshellgetcontext_        STSHELLGETCONTEXT
-#define stshellsetapply_          STSHELLSETAPPLY
-#define stshellsetapplytranspose_ STSHELLSETAPPLYTRANSPOSE
-#define stshellsetbacktransform_  STSHELLSETBACKTRANSFORM
+#define stshellgetcontext_                 STSHELLGETCONTEXT
+#define stshellsetapply_                   STSHELLSETAPPLY
+#define stshellsetapplytranspose_          STSHELLSETAPPLYTRANSPOSE
+#define stshellsetapplyhermitiantranspose_ STSHELLSETAPPLYHERMITIANTRANSPOSE
+#define stshellsetbacktransform_           STSHELLSETBACKTRANSFORM
 #elif !defined(PETSC_HAVE_FORTRAN_UNDERSCORE)
-#define stshellgetcontext_        stshellgetcontext
-#define stshellsetapply_          stshellsetapply
-#define stshellsetapplytranspose_ stshellsetapplytranspose
-#define stshellsetbacktransform_  stshellsetbacktransform
+#define stshellgetcontext_                 stshellgetcontext
+#define stshellsetapply_                   stshellsetapply
+#define stshellsetapplytranspose_          stshellsetapplytranspose
+#define stshellsetapplyhermitiantranspose_ stshellsetapplyhermitiantranspose
+#define stshellsetbacktransform_           stshellsetbacktransform
 #endif
 
 static struct {
   PetscFortranCallbackId apply;
   PetscFortranCallbackId applytranspose;
+  PetscFortranCallbackId applyhermtrans;
   PetscFortranCallbackId backtransform;
 } _cb;
 
@@ -38,6 +41,11 @@ static PetscErrorCode ourshellapply(ST st,Vec x,Vec y)
 static PetscErrorCode ourshellapplytranspose(ST st,Vec x,Vec y)
 {
   PetscObjectUseFortranCallback(st,_cb.applytranspose,(ST*,Vec*,Vec*,PetscErrorCode*),(&st,&x,&y,&ierr));
+}
+
+static PetscErrorCode ourshellapplyhermitiantranspose(ST st,Vec x,Vec y)
+{
+  PetscObjectUseFortranCallback(st,_cb.applyhermtrans,(ST*,Vec*,Vec*,PetscErrorCode*),(&st,&x,&y,&ierr));
 }
 
 static PetscErrorCode ourshellbacktransform(ST st,PetscInt n,PetscScalar *eigr,PetscScalar *eigi)
@@ -60,6 +68,12 @@ SLEPC_EXTERN void stshellsetapplytranspose_(ST *st,void (*applytranspose)(void*,
 {
   *ierr = PetscObjectSetFortranCallback((PetscObject)*st,PETSC_FORTRAN_CALLBACK_CLASS,&_cb.applytranspose,(PetscVoidFunction)applytranspose,NULL); if (*ierr) return;
   *ierr = STShellSetApplyTranspose(*st,ourshellapplytranspose);
+}
+
+SLEPC_EXTERN void stshellsetapplyhermitiantranspose_(ST *st,void (*applyhermtrans)(void*,Vec*,Vec*,PetscErrorCode*),PetscErrorCode *ierr)
+{
+  *ierr = PetscObjectSetFortranCallback((PetscObject)*st,PETSC_FORTRAN_CALLBACK_CLASS,&_cb.applyhermtrans,(PetscVoidFunction)applyhermtrans,NULL); if (*ierr) return;
+  *ierr = STShellSetApplyHermitianTranspose(*st,ourshellapplyhermitiantranspose);
 }
 
 SLEPC_EXTERN void stshellsetbacktransform_(ST *st,void (*backtransform)(void*,PetscScalar*,PetscScalar*,PetscErrorCode*),PetscErrorCode *ierr)