+++ src/sys/classes/st/impls/shift/shift.c
@@ -130,6 +130,7 @@ SLEPC_EXTERN PetscErrorCode STCreate_Shift(ST st)
 
   st->ops->apply           = STApply_Generic;
   st->ops->applytrans      = STApplyTranspose_Generic;
+  st->ops->applyhermtrans  = STApplyHermitianTranspose_Generic;
   st->ops->backtransform   = STBackTransform_Shift;
   st->ops->setshift        = STSetShift_Shift;
   st->ops->getbilinearform = STGetBilinearForm_Default;