+++ src/sys/classes/st/tests/test4.c
@@ -166,12 +166,10 @@ int main(int argc,char **argv)
    test:
       suffix: 1
       args: -st_transform -st_matmode {{copy shell}}
-      output_file: output/test4_1.out
       requires: !single
 
    test:
       suffix: 2
       args: -st_matmode {{copy shell}}
-      output_file: output/test4_2.out
 
 TEST*/